Refactor TrieTxn
Rename TrieTxn to TransactionStorage and move it to the core package.
