+++ core/transaction_storage_test.go
@@ -1,10 +1,11 @@
-package trie
+package core
 
 import (
 	"errors"
 	"testing"
 
 	"github.com/NethermindEth/juno/core/felt"
+	"github.com/NethermindEth/juno/core/trie"
 	"github.com/NethermindEth/juno/db"
 	"github.com/NethermindEth/juno/db/pebble"
 	"github.com/bits-and-blooms/bitset"
@@ -16,19 +17,19 @@ func TestTrieTxn(t *testing.T) {
 	prefix := []byte{37, 44}
 
 	key := bitset.New(44)
-	node := new(Node)
+	node := new(trie.Node)
 	node.Value, _ = new(felt.Felt).SetRandom()
 
 	// put a node
 	assert.NoError(t, testDb.Update(func(txn db.Transaction) error {
-		tTxn := &TrieTxn{txn, prefix}
+		tTxn := &TransactionStorage{txn, prefix}
 
 		return tTxn.Put(key, node)
 	}))
 
 	// get node
 	assert.NoError(t, testDb.View(func(txn db.Transaction) error {
-		tTxn := &TrieTxn{txn, prefix}
+		tTxn := &TransactionStorage{txn, prefix}
 
 		got, err := tTxn.Get(key)
 		assert.NoError(t, err)
@@ -39,7 +40,7 @@ func TestTrieTxn(t *testing.T) {
 
 	// in case of an error, tx should roll back
 	assert.Error(t, testDb.Update(func(txn db.Transaction) error {
-		tTxn := &TrieTxn{txn, prefix}
+		tTxn := &TransactionStorage{txn, prefix}
 
 		if err := tTxn.Delete(key); err != nil {
 			t.Error(err)
@@ -50,7 +51,7 @@ func TestTrieTxn(t *testing.T) {
 
 	// should still be able to get the node
 	assert.NoError(t, testDb.View(func(txn db.Transaction) error {
-		tTxn := &TrieTxn{txn, prefix}
+		tTxn := &TransactionStorage{txn, prefix}
 
 		got, err := tTxn.Get(key)
 		assert.Equal(t, true, got.Equal(node))
@@ -60,13 +61,13 @@ func TestTrieTxn(t *testing.T) {
 
 	// successful delete
 	assert.NoError(t, testDb.Update(func(txn db.Transaction) error {
-		tTxn := &TrieTxn{txn, prefix}
+		tTxn := &TransactionStorage{txn, prefix}
 		return tTxn.Delete(key)
 	}))
 
 	// should error with key not found
 	assert.EqualError(t, testDb.View(func(txn db.Transaction) error {
-		tTxn := &TrieTxn{txn, prefix}
+		tTxn := &TransactionStorage{txn, prefix}
 		_, err := tTxn.Get(key)
 		return err
 	}), "Key not found")
